Rename SetStrokeGeometry to SetPolylineGeometry
